pyproject.toml
[project]
name = "sales-rag"
version = "0.1.0"
description = "OpenClaw sales analytics agent — Olist Brazilian E-Commerce dataset"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "kaggle>=2.0.0",
    "pandas>=2.2.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=13.0",
    "cohere>=5.0.0",
    "chromadb>=1.5.7",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501", "E402"]
